chore(skills): move `ForceShield::execution` into traits
